Cleanups, refactorings, fixes
