+++ src/main/java/com/teamwizardry/wizardry/common/core/version/manifest/ManifestHandler.java
@@ -1,14 +1,6 @@
-package com.teamwizardry.wizardry.common.core.version;
+package com.teamwizardry.wizardry.common.core.version.manifest;
 
 import com.google.common.io.Files;
-import com.google.gson.JsonArray;
-import com.google.gson.JsonElement;
-import com.google.gson.JsonIOException;
-import com.google.gson.JsonObject;
-import com.google.gson.JsonParser;
-import com.google.gson.JsonSyntaxException;
-import com.google.gson.internal.Streams;
-import com.google.gson.stream.JsonWriter;
 import com.teamwizardry.librarianlib.core.LibrarianLib;
 import com.teamwizardry.wizardry.Wizardry;
 import com.teamwizardry.wizardry.crafting.mana.ManaRecipes;
@@ -95,7 +87,7 @@ public class ManifestHandler {
 					e.printStackTrace();
 				}
 			}
-			writeJsonToFile(generateManifestJson(internalManifestMap), externalManifest);
+			ManifestUtils.writeJsonToFile(ManifestUtils.generateManifestJson(internalManifestMap), externalManifest);
 			externalManifestMap.putAll(internalManifestMap);
 			Wizardry.logger.info("    > Successfully generated new manifest file");
 		}
@@ -179,7 +171,7 @@ public class ManifestHandler {
 					return;
 				}
 
-				writeJsonToFile(generateManifestJson(internalManifestMap), externalManifest);
+				ManifestUtils.writeJsonToFile(ManifestUtils.generateManifestJson(internalManifestMap), externalManifest);
 				externalManifestMap.putAll(internalManifestMap);
 				generatedNewManifest = true;
 				Wizardry.logger.info("    > Successfully generated new manifest file");
@@ -192,7 +184,7 @@ public class ManifestHandler {
 			}
 
 			Wizardry.logger.info("    > Found manifest file. Reading...");
-			loadManifestFile(externalManifest, externalManifestMap, true);
+			ManifestUtils.loadManifestFile(externalManifest, externalManifestMap, true);
 			Wizardry.logger.info("    >  |____________________________________/");
 
 		} catch (IOException e) {
@@ -200,42 +192,6 @@ public class ManifestHandler {
 		}
 	}
 	
-	static void loadManifestFile(File externalManifest, HashMap<String, HashMap<String, String>> manifestMap, boolean verbose) throws IOException {
-		JsonElement element = new JsonParser().parse(new FileReader(externalManifest));
-
-		if (element != null && element.isJsonObject()) {
-
-			for (Map.Entry<String, JsonElement> categorySet : element.getAsJsonObject().entrySet()) {
-				String category = categorySet.getKey();
-				JsonElement categoryElement = categorySet.getValue();
-
-				manifestMap.putIfAbsent(category, new HashMap<>());
-				if( verbose ) {
-					Wizardry.logger.info("    >  |");
-					Wizardry.logger.info("    >  |_ Category found: " + category);
-				}
-
-				if (categoryElement.isJsonArray()) {
-					for (JsonElement element1 : categoryElement.getAsJsonArray()) {
-						if (!element1.isJsonObject()) continue;
-
-						JsonObject externalObject = element1.getAsJsonObject();
-
-						if (!externalObject.has("id") || !externalObject.has("hash")) continue;
-
-						String id = externalObject.getAsJsonPrimitive("id").getAsString();
-						String hash = externalObject.getAsJsonPrimitive("hash").getAsString();
-
-						manifestMap.get(category).put(id, hash);
-						if( verbose ) {
-							Wizardry.logger.info("    >  | |_ " + id + ": " + hash);
-						}
-					}
-				}
-			}
-		}
-	}
-
 	public void addItemToManifest(String category, String modId, String id, File file) {
 		internalManifestMap.putIfAbsent(category, new HashMap<>());
 
@@ -260,35 +216,5 @@ public class ManifestHandler {
 		String prevModId = fileToMod.get(category).put(id, modId);
 		if( prevModId != null )
 			Wizardry.logger.warn("    > File name conflict for " + category + "/" + id + ".json occurring in mods '" + modId + "' and '" + prevModId + "'. Some stuff wont be available." );
-	}
-
-	static JsonObject generateManifestJson(HashMap<String, HashMap<String, String>> manifestMap) {
-		JsonObject jsonManifest = new JsonObject();
-
-		for (Map.Entry<String, HashMap<String, String>> categoryEntry : manifestMap.entrySet()) {
-			String category = categoryEntry.getKey();
-
-			JsonArray categoryArray = new JsonArray();
-			for (Map.Entry<String, String> entry : categoryEntry.getValue().entrySet()) {
-				JsonObject entryObject = new JsonObject();
-				entryObject.addProperty("id", entry.getKey());
-				entryObject.addProperty("hash", entry.getValue());
-
-				categoryArray.add(entryObject);
-			}
-
-			jsonManifest.add(category, categoryArray);
-		}
-
-		return jsonManifest;
-	}
-
-	public static void writeJsonToFile(JsonObject object, File file) {
-		try (JsonWriter writer = new JsonWriter(Files.newWriter(file, Charset.defaultCharset()))) {
-			Streams.write(object, writer);
-		} catch (IOException e) {
-			Wizardry.logger.error("    > SOMETHING WENT WRONG! Could not create or write to file! Customizations to recipes and modules will be reset every time you load the game!");
-			e.printStackTrace();
-		}
 	}
 }
