+++ src/main/java/com/teamwizardry/wizardry/proxy/CommonProxy.java
@@ -14,8 +14,8 @@ import com.teamwizardry.wizardry.client.gui.GuiHandler;
 import com.teamwizardry.wizardry.client.gui.book.PageWizardryStructure;
 import com.teamwizardry.wizardry.common.advancement.AchievementEvents;
 import com.teamwizardry.wizardry.common.core.EventHandler;
-import com.teamwizardry.wizardry.common.core.version.ManifestHandler;
-import com.teamwizardry.wizardry.common.core.version.ManifestUpgrader;
+import com.teamwizardry.wizardry.common.core.version.manifest.ManifestHandler;
+import com.teamwizardry.wizardry.common.core.version.manifest.ManifestUpgrader;
 import com.teamwizardry.wizardry.common.item.ItemBook;
 import com.teamwizardry.wizardry.common.module.effects.ModuleEffectLeap;
 import com.teamwizardry.wizardry.common.module.effects.ModuleEffectTimeSlow;
